Restructure unshard.py into unshard/ package with Pydantic models
- types.py: AxisInfo, ConcatParams (discriminated union ready),
  UnshardStep, UnshardPlan (with pick_world_ranks)
- parallel_info.py: normalize_parallel_info
- plan.py: compute_unshard_plan + _resolve_unshard_params
- execute.py: execute_unshard_plan + unshard_concat
- entrypoint.py: use list indices consistently, remove world_ranks hack
