+++ python/sglang/srt/debug_utils/comparator/unshard/__init__.py
@@ -1,0 +1,25 @@
+from sglang.srt.debug_utils.comparator.unshard.execute import (
+    execute_unshard_plan,
+    unshard_concat,
+)
+from sglang.srt.debug_utils.comparator.unshard.parallel_info import (
+    normalize_parallel_info,
+)
+from sglang.srt.debug_utils.comparator.unshard.plan import compute_unshard_plan
+from sglang.srt.debug_utils.comparator.unshard.types import (
+    AxisInfo,
+    ConcatParams,
+    UnshardPlan,
+    UnshardStep,
+)
+
+__all__ = [
+    "AxisInfo",
+    "ConcatParams",
+    "UnshardPlan",
+    "UnshardStep",
+    "compute_unshard_plan",
+    "execute_unshard_plan",
+    "normalize_parallel_info",
+    "unshard_concat",
+]
